add onyx request id
