+++ backend/model_server/main.py
@@ -19,10 +19,8 @@ from model_server.encoders import router as encoders_router
 from model_server.management_endpoints import router as management_router
 from model_server.utils import get_gpu_type
 from onyx import __version__
-from onyx.server.middleware.request_id_middleware import (
-    add_fastapi_request_id_middleware,
-)
 from onyx.utils.logger import setup_logger
+from onyx.utils.middleware import add_onyx_request_id_middleware
 from shared_configs.configs import INDEXING_ONLY
 from shared_configs.configs import MIN_THREADS_ML_MODELS
 from shared_configs.configs import MODEL_SERVER_ALLOWED_HOST
@@ -115,7 +113,7 @@ def get_model_app() -> FastAPI:
     application.include_router(encoders_router)
     application.include_router(custom_models_router)
 
-    add_fastapi_request_id_middleware(application, "MDL", logger)
+    add_onyx_request_id_middleware(application, "MDL", logger)
 
     return application
